Add files via upload
This is the OpenLog, a simple yet powerful serial data logger.  Based on an ATmega328, it automatically logs any serial data it receives to a microSD card. It supports baud rates up to 115200bps and is configured via a simple text file on the card, making it an essential tool for capturing data from sensors or debugging remote devices.